refactor sub screens to respective folders
initiate messaging feature
