+++ src/screens/settings/Settings.js
@@ -8,7 +8,7 @@ import {
 } from "react-native-rapi-ui";
 import { Ionicons } from "@expo/vector-icons";
 import { getAuth, signOut } from "firebase/auth"; // Assuming you are using Firebase Auth
-import CustomTopNav from "../components/CustomTopNav";
+import CustomTopNav from "../../components/CustomTopNav";
 
 export default function ({ navigation }) {
   const { isDarkmode, setTheme } = useTheme();
@@ -26,7 +26,7 @@ export default function ({ navigation }) {
   return (
     <Layout>
       <CustomTopNav title="Settings" navigation={navigation} />
-      <View style={{...styles.container, backgroundColor: isDarkmode ? themeColor.dark : themeColor.white}}>
+      <View style={{ ...styles.container, backgroundColor: isDarkmode ? themeColor.dark : themeColor.white }}>
         <TouchableOpacity style={{ ...styles.option, borderBottomColor: isDarkmode ? themeColor.white200 : themeColor.dark }} onPress={() => navigation.navigate('Profile')}>
           <Text>Profile</Text>
           <Ionicons name="chevron-forward" size={20} color={isDarkmode ? themeColor.white200 : themeColor.dark} />
